+++ src/main/webapp/jsp/article/detail.jsp
@@ -5,6 +5,7 @@
     
 <%
 	Map<String, Object> todo = (Map<String, Object>) request.getAttribute("articleMap");
+	int loginedMemberId = (int) request.getAttribute("loginedMemberId");
 %>
 <!DOCTYPE html>
 <html>
@@ -18,11 +19,14 @@
 		<div>번호 : <%= (int) todo.get("id") %></div>
 		<div>작성일 : <%= (LocalDateTime) todo.get("regDate") %></div>
 		<div>수정일 : <%= (LocalDateTime) todo.get("updateDate") %></div>
+		<div>작성자 : <%= (String) todo.get("name") %></div>
 		<div>제 목 : <%= (String) todo.get("title") %></div>
 		<div style="white-space: pre-wrap;">내 용 : <%= (String )todo.get("body") %></div>
 	</div>
-	<a href="modify?id=<%=(int)todo.get("id")%>">게시물 수정</a>
-	<a href="delete?id=<%=(int)todo.get("id")%>" onclick="return confirm('게시물을 삭제하시겠습니까?')">게시물 삭제</a>
+	<%if(loginedMemberId != -1 && loginedMemberId == (int) todo.get("memberId")) {%>
+		<a href="modify?id=<%=(int)todo.get("id")%>">게시물 수정</a>
+		<a href="delete?id=<%=(int)todo.get("id")%>" onclick="return confirm('게시물을 삭제하시겠습니까?')">게시물 삭제</a>
+	<%} %>
 	<br />
 	<br />
 	<a href="list">게시물 목록</a>
